requirements.txt
amqp==2.6.0
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
billiard==3.5.0.5
bleach==3.1.5
cachelib==0.1.1
celery==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
chart-studio==1.1.0
click==7.1.2
colorama==0.4.3
cs50==5.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==3.0.8
django-admin-bootstrapped==2.5.4
django-appconf==1.0.4
django-autoslug-iplweb==1.9.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-classy-tags==1.0.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-easy-pdf==0.1.0
django-imagekit==3.2.7
django-markdownx==3.0.1
django-phonenumber-field==1.3.0
django-rest-authtoken==1.2.4
django-rest-framework==0.1.0
django-simple-history==1.8.1
django-taggit==1.2.0
django-taggit-templatetags==0.2.5
django-taggit-templatetags2==1.6.1
django-templatetag-sugar==1.0
django-widget-tweaks==1.4.1
django-widgets==0.1.16
django-widgets-improved==1.5.0
djangorestframework==3.11.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
html5lib==1.1
idna==2.9
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
kombu==4.6.11
lxml==4.5.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
mysql-connector-python==8.0.20
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.0
packaging==20.4
pandas==1.0.5
pandas-alive==0.2.2
pandocfilters==1.4.2
parso==0.7.0
passlib==1.7.2
phonenumberslite==8.12.6
pickleshare==0.7.5
pilkit==1.1.13
Pillow==7.1.2
pipenv==2020.6.2
playsound==1.2.2
plotly==4.8.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
pygame==1.9.6
Pygments==2.6.1
PyMySQL==0.9.3
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
reportlab==3.5.44
requests==2.24.0
retrying==1.3.3
Send2Trash==1.5.0
six==1.9.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
sqlparse==0.3.1
static3==0.6.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.28
virtualenv-clone==0.5.4
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
WTForms==2.3.1
xhtml2pdf==0.2.4
mysqlclient==1.4.4